Added share object to the simulation.rs test and fixes everything that needed to work.

+++ src/entity/file_blocks.rs
@@ -5,7 +5,7 @@ use sea_orm::entity::prelude::*;
 #[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq)]
 #[sea_orm(table_name = "file_blocks")]
 pub struct Model {
-	pub file_id: i32,
+	pub file_id: i64,
 	#[sea_orm(primary_key, auto_increment = false)]
 	pub block_hash: String,
 	pub sequence: i32,
